fix stuff

+++ src/map/status.cpp
@@ -128,7 +128,7 @@ uint64 SizeFixDatabase::parseBodyNode(const YAML::Node &node) {
 	if (!this->asString(node, "Weapon", weapon_name))
 		return 0;
 
-	int weapon_id;
+	int64 weapon_id;
 
 	if (!script_get_constant(weapon_name.c_str(), &weapon_id)) {
 		this->invalidWarning(node, "Size Fix unknown weapon %s, skipping.\n", weapon_name.c_str());
